fix the warning message while compile

+++ src/endstop.c
@@ -10,7 +10,7 @@
 #include "command.h" // DECL_COMMAND
 #include "sched.h" // struct timer
 #include "trsync.h" // trsync_do_trigger
-
+#include "BD_sensor.h"
 struct endstop {
     struct timer time;
     struct gpio_in pin;
@@ -29,9 +29,8 @@ uint8_t read_endstop_pin(struct endstop *e)
 	uint8_t state_e=0;
 	if(e->type==2)
 	{
-		uint32_t tm=Get_Distane_data();
+		uint16_t tm=Get_Distane_data();
 		state_e=(tm>=0.01)?0:1;
-
 	}
 	else	
 		state_e=gpio_in_read(e->pin);
